requirements.txt
asgiref==3.3.1
bfa==1.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
idna==2.10
mysql-connector==2.2.9
Pillow==8.1.2
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
sweetify==2.0.0
urllib3==1.26.3
Werkzeug==1.0.1
